created new builtin method "includeJs" (available for javascript only)
- files marked as "useAsJavaScriptLibrary" will not automatically be prepended in **every** JavaScript context anymore
- upon calling "Structr.includeJs('filename.js')" the contents of that specific JavaScript file will be interpreted and inserted in the current context
  therefore it may be wise to make these calls at the start of a javascript context where the files are needed.

Before this change the JavaScript files were put in the global scope and could therefore not interfere with the code in the scope. The different files could however interfere with each other.
After this change the code will be inserted in the current scope. This means that the code can interfere with each other AND the current scope.

It is highly recommended that you namespace/scope the code in these JavaScript Library files in order to reduce possible interferences.